add xcm transfer verification before submission
