+++ novawallet/Common/Services/ExtrinsicService/Substrate/Xcm/XcmTransferService+Protocol.swift
@@ -118,12 +118,26 @@ extension XcmTransferService: XcmTransferServiceProtocol {
                 chainAccount: chainAccount
             )
 
+            let verificationWrapper = submissionVerifier.createVerificationWrapper(
+                for: request.unweighted,
+                callOrigin: .system(.signed(chainAccount.accountId)),
+                callClosure: {
+                    try callBuilderWrapper.targetOperation.extractNoCancellableResultData()
+                }
+            )
+
+            verificationWrapper.addDependency(wrapper: callBuilderWrapper)
+
             let submitWrapper = operationFactory.submit({ builder in
+                // submit extrinsic only if verification passed
+                try verificationWrapper.targetOperation.extractNoCancellableResultData()
+
                 let collector = try callBuilderWrapper.targetOperation.extractNoCancellableResultData()
                 return try collector.addingToExtrinsic(builder: builder)
             }, signer: signer, payingIn: request.originFeeAsset)
 
             submitWrapper.addDependency(wrapper: callBuilderWrapper)
+            submitWrapper.addDependency(wrapper: verificationWrapper)
 
             submitWrapper.targetOperation.completionBlock = {
                 do {
@@ -140,7 +154,8 @@ extension XcmTransferService: XcmTransferServiceProtocol {
                 }
             }
 
-            let operations = callBuilderWrapper.allOperations + submitWrapper.allOperations
+            let operations = callBuilderWrapper.allOperations + verificationWrapper.allOperations +
+                submitWrapper.allOperations
 
             operationQueue.addOperations(operations, waitUntilFinished: false)
         } catch {
