+++ novawallet/Common/Substrate/Calls/Builder/RuntimeCallCollecting+AnyRuntimeCall.swift
@@ -1,0 +1,9 @@
+import Foundation
+import SubstrateSdk
+
+extension RuntimeCallCollecting {
+    func toAnyRuntimeCall(with context: RuntimeJsonContext) throws -> AnyRuntimeCall {
+        let builder = RuntimeCallBuilder(context: context)
+        return try addingToCall(builder: builder, toEnd: true).build()
+    }
+}
